--- src/app/(client)/driver-registration/components/PhoneNumber/types.ts
@@ -1,5 +1,0 @@
-import { DriverRegistrationFormValues } from '@/app/(client)/driver-registration/components/Form/schema';
-
-export type PhoneNumberProps = {
-  onSubmit: (values: DriverRegistrationFormValues) => void;
-};
