+++ src/app/(client)/driver-registration/components/ProgressBar/types.ts
@@ -1,3 +1,5 @@
+import { DriverRegistrationStep } from '@/app/(client)/driver-registration/constants';
+
 export type ProgressBarProps = {
-  step: string;
+  step: DriverRegistrationStep;
 };
